backendapi/requirements.txt
asgiref
Django
django-cors-headers
django-debug-toolbar
django-filter
django-versatileimagefield
djangorestframework
djangorestframework-simplejwt
drf-flex-fields
importlib-metadata
Markdown
Pillow
PyJWT
python-magic
python-magic-bin
pytz
six
sqlparse
typing
zipp
